Don't open file twice when specified by absolute path.
When invoking `require '/a.rb'` (i.e. via an absolute path), ruby
generates this sequence of syscalls:

    open    /a.rb
    fstat64 /a.rb
    close   /a.rb
    open    /a.rb
    fstat64 /a.rb
    fstat64 /a.rb
    read    /a.rb
    close   /a.rb

It is apparent that the only inherently necessary members of this
sequence are:

    open    /a.rb
    fstat64 /a.rb
    read    /a.rb
    close   /a.rb

(the fstat64 isn't *obviously* necessary, but it does serve a purpose
and probably shouldn't be removed).

The first open/fstat64/close is used to check whether the file is
loadable. This is important when scanning the `$LOAD_PATH`, since it is
used to determine when a file has been found. However, when we've
already unambiguously identified a file before invoking `require`, this
serves no inherent purpose, since we can move whatever work is happening
as a result of that `fstat64` into the second open/close sequence.

This change bypasses the first open/fstat64/close in the case of an
absolute path to `require`. It also removes one of the doubled-up
`fstat64` calls later in the sequence. As a result, the number of
syscalls to require a file changes:

* From 8 to 4 when specified by absolute path;
* From 8 to 7 otherwise.

In future work, it would be possible to re-use the file descriptor
opened while searching the `$LOAD_PATH` without the close/open sequence,
but this would cause some ugly layering issues.